Refactor and enhance API handling
- Added multiple NuGet packages and project references to MRMDesktopUI.Library for extended functionality.
- Reorganized APIHelper and AuthenticatedUser classes into MRMDesktopUI.Library for better modularity.
- Refactored namespaces for APIHelper and IAPIHelper to MRMDesktopUI.Library.Api, aligning with project structure.
- Updated MRMDesktopUI project to reference the updated library project and namespaces.
- Configured assembly binding redirects for Newtonsoft.Json to version 13.0.0.0 to avoid conflicts.
- Implemented new functionality in APIHelper for initializing HttpClient and user authentication.
- Defined IAPIHelper interface to standardize API interaction methods.
- Removed redundant files from MRMDesktopUI project post-reorganization.
