+++ MRMDesktopUI.Library/Models/AuthenticatedUser.cs
